Add ViewStudent class
